refactor command issuer
refactor: custom actions structure
